+++ CHANGELOG.md
@@ -21,7 +21,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- upgraded Flutter to stable v3.3.3
+- upgraded Flutter to stable v3.3.4
 
 ### Fixed